requirements.txt
six
msgpack-python
PyYAML
